feat(): module automation - add widgets programatically
